+++ src/util/EIOCompat.ts
