+++ src/util/utf8ReadMonkeypatch.ts
